恢复detail面板
